bpf: Fix sk_skb_adjust_room full_len update
When we generate replies from a verdict program and try to resize the
packet to fit the reponse in the payload, the response gets dropped by
the strparser layer. The reason was that the skb->len doesn't add up to
the strp_msg full_len. Ensure the full_len is kept up to date whenever
the helper is called to make sure the response is correctly generated.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>