bpf: Add bpf_register_heap kfunc
We are going to introduce support for a BTF type tag on pointer types,
called __hptr(heap), where heap refers to the name of a map that serves
as the allocation context of the pointer.

It might be the case that the program never really refers to the map at
all, instead always interacting to its values through the __hptr. In
such a case, we need to be able to have the map in our used_maps array
for the program during verification to lookup the heap by its name in
the BTF type tag.

Introduce a kfunc bpf_register_heap, which is a verifier builtin, i.e.
it only takes effect during verification and is replaced by a NOP
instruction at runtime during JIT. A user must any heaps used by the
program using this kfunc so that the name lookup can work correctly. We
can improve this mechanism in the future to be automatic, but for now,
this manual registration is required.

TODO(kkd): It might make sense for libbpf to insert pseudo load
instructions for the maps so that they get picked up by the verifier.
That might be cheaper than doing this trick in the verifier.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>