bpf: Introduce bpf_preempt_[disable,enable]
Introduce kfuncs to manage preemption context within a BPF program.
There are certain key things we need to keep in mind: ensure that
the preempt disable section isn't allowed to be unecessarily long,
make sure that no sleepable kernel code can be invoked while we are
in such a context, and ensure that memory allocating helpers do the
GFP_ATOMIC equivalent for kmalloc when sleepable programs invoke them in
a preempt disabled region.

Other than that, the code matches RCU kfuncs and spin lock helpers
protection mechanisms around the verifier ensuring that they are
correctly matched and preemption is enabled before the program is
allowed to exit.

Ensure that we also do not allow bpf_throw to bypass these restrictions,
and put the check to ensure unmatched preempt_disable within the block
that is reachable by process_bpf_exit_full.

This helper should allow us to manage per-CPU data structures from
within a BPF program. The equivalent can be done using rseq in
userspace, but sometimes we have use-cases of kernel-only data
structures as well.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>