bpf: Temporarily enable dynptr_slice{,_rdwr} in spin lock section [TODO(kkd): Fix]
Let's allow programs to temporarily call bpf_dynptr_slice{,_rdwr} in a
spin lock critical section. This should be a constant time operation
given an input and should not be taking too much time, neither does it
take any locks that might deadlock with the bpf_spin_lock. Hence,
temporarily remove the restriction on it, and deal with a proper fix
later.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>